+++ ui/src/components/Colors.jsx
@@ -1,0 +1,12 @@
+export const colorMap = {
+    'OTS': 'rgb(30,185,128)',
+    'ROR': '#dc004e',
+    'DOR': '#82ca9d',
+    'TOS': '#ff7300',
+    'COI': '#8884d8',
+    'EDO': '#8dd1e1',
+    'ROS': '#ffc658',
+    'ROD': '#d0ed57',
+    'Overview': 'rgb(255, 255, 255)',
+    'defaultColor': 'rgb(255, 255, 255)'
+};
