+++ ui/src/styles/main.css
@@ -19,3 +19,4 @@ a:hover, a:active, a:focus {
     color: #fff;
     outline: none;
 }
+
